move getLocation to Context
