+++ frontend/src/components/ListingPage/ListingPage.jsx
@@ -4,29 +4,32 @@ import { useEffect, useState } from "react";
 import { useLocation } from 'react-router-dom';
 import axios from "axios";
 import "./ListingPage.css";
+import { useLocationContext } from "../../context/Location/LocationContext";
+import { CircularProgress } from "@mui/material";
 
-export const ListingPage = ({ listing }) => {
+export const ListingPage = () => {
   const [listings, setListings] = useState([])
   const location = useLocation();
+  const { address } = useLocationContext();
 
   useEffect(() => {
     const getListings = async () => {
       try {
         let res = await axios.get("http://localhost:8000/listing" + location.search);
         setListings(res.data);
-        } catch (error) {
-            console.log(error);
-        }
+      } catch (error) {
+        console.log(error);
+      }
     }
     getListings();
   }, [])
 
   return (
     <div className="listing-page">
-      <UserPanel></UserPanel>
-      {listings && listings.map((listing, key) => (
-        <ListingPanel listing={listing} key={key}/>
-      ))}
+      {address ? <h4>Viewing listings around {address?.split(",")[0]}</h4> : <CircularProgress />}
+      {listings.length ? listings.map((listing, key) => (
+        <ListingPanel listing={listing} key={key} />
+      )) : <CircularProgress />}
     </div>
   );
 };
