+++ frontend/src/index.js
@@ -3,10 +3,13 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import { AuthProvider } from "./context/Auth/AuthContext";
+import { LocationProvider } from "./context/Location/LocationContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <AuthProvider>
-    <App />
+    <LocationProvider>
+       <App />
+    </LocationProvider>
   </AuthProvider>
 );
